VR hands system and CCTV of robot
features:
Aim to the robot where you want it to stand.
Gesture to the robot to come to you.
Control the Robot Throught Montotr
Seperate palyer control for the robot and thep player.
Robot can Double jumps
Robot Jump can be controlled and modified to become faster higher and or slower.
Added assets for visuals purpose.